[MAS19BLA-45] Poprawki po pr #time 5min
