+++ frontend/src/app/views/evaluations/evaluations.component.ts
@@ -43,47 +43,4 @@ export class EvaluationsComponent implements OnInit {
     }
     return sum;
   }
-
-  @HostListener('document:keyup', ['$event'])
-  onKeyUp(event: KeyboardEvent) {
-    if (this.userService.isCandidate()) {
-      if (event.key === 'PrintScreen') {
-        this.copyToClipboard();
-        event.preventDefault();
-      }
-    }
-  }
-
-  @HostListener('document:contextmenu', ['$event'])
-  onClick1($event) {
-    if (this.userService.isCandidate()) {
-      event.preventDefault();
-    }
-  }
-
-  @HostListener('mouseup', ['$event']) onClick($event) {
-    if (this.userService.isCandidate()) {
-      if ($event.which === 3) {
-        this.toastr.error('Nie wolno klikać prawym przyciskiem myszy!', 'Drogi kandydacie', {
-          timeOut: 3000,
-          closeButton: true,
-        });
-        $event.preventDefault();
-      }
-    }
-  }
-
-
-  copyToClipboard() {
-    const aux = document.createElement('input');
-    aux.setAttribute('value', 'print screen disabled!');
-    document.body.appendChild(aux);
-    aux.select();
-    document.execCommand('copy');
-    document.body.removeChild(aux);
-    this.toastr.error('Nie wolno robić print screen!', 'Drogi kandydacie', {
-      timeOut: 3000,
-      closeButton: true,
-    });
-  }
 }
