+++ frontend/src/app/views/solve-test-form/solve-test-form.component.ts
@@ -64,42 +64,4 @@ export class SolveTestFormComponent implements OnInit {
     await this.testAnswerService.create(testAnswer);
     await this.router.navigate(['/tests']);
   }
-
-  @HostListener('document:keyup', ['$event'])
-  onKeyUp(event: KeyboardEvent) {
-    if (event.key === 'PrintScreen') {
-      this.copyToClipboard();
-      event.preventDefault();
-    }
-  }
-
-  @HostListener('document:contextmenu', ['$event'])
-  onClick1($event) {
-    event.preventDefault();
-  }
-
-  @HostListener('mouseup', ['$event']) onClick($event) {
-
-    if ($event.which === 3) {
-      this.toastr.error('Nie wolno klikać prawym przyciskiem myszy!', 'Drogi kandydacie', {
-        timeOut: 3000,
-        closeButton: true,
-      });
-      $event.preventDefault();
-    }
-  }
-
-
-  copyToClipboard() {
-    const aux = document.createElement('input');
-    aux.setAttribute('value', 'print screen disabled!');
-    document.body.appendChild(aux);
-    aux.select();
-    document.execCommand('copy');
-    document.body.removeChild(aux);
-    this.toastr.error('Nie wolno robić print screen!', 'Drogi kandydacie', {
-      timeOut: 3000,
-      closeButton: true,
-    });
-  }
 }
